feat: implement sound on off
